EsalPlatform/apps/api/requirements.txt
# ESAL Platform Backend Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Validation & Serialization
pydantic==2.5.1
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0
PyJWT==2.8.0

# Supabase Integration
supabase==2.3.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# AI Integration
google-generativeai==0.3.2

# Environment Variables
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
